refactor(copilot): add pyproject.toml + poetry, rename modules, move tests #66
Python packaging for TDD workflow:
- Add pyproject.toml (package-mode=false) with deps + pytest config
- Add poetry.lock for reproducible builds
- Rename _browser.py -> browser.py (drop _ prefix from all modules)
- Move *_test.py -> tests/test_*.py (standard test layout)
- Add tests/conftest.py with shared fixtures
- Update all imports to match renamed modules
- Update spec references (_browser.py -> browser.py)
- Fix z85 dep: use pyzmq (z85 via zmq.utils.z85)
- 121 Python tests pass, 9 Go tests pass

Developer workflow: cd tools/copilot && poetry install && poetry run pytest
