+++ specs/COPILOT-53-TO-80.md
@@ -44,7 +44,7 @@ No tooling changes needed — just re-run with correct information.
 | Mass Dispel | 1 | Click X on 5+ notifications simultaneously | Notifications auto-dismiss in ~7s |
 | Privacy Policy | 3 | Find hidden element in Angular page | Need to render and inspect Angular bindings |
 
-**Fix: Add `browser open_tab` action (~50 LoC in `_browser.py`).**
+**Fix: Add `browser open_tab` action (~50 LoC in `browser.py`).**
 
 ### C. Missing Crypto/Encoding Tools (5 challenges)
