+++ specs/COPILOT.md
@@ -15,12 +15,12 @@ Browser-driven, LLM-reasoned pentesting via a turn-based reasoning loop.
 │  Container         │                                          │
 │  ┌─────────────────▼──────────────────┐                      │
 │  │  loop.py (JSON-RPC server)         │                      │
-│  │  ├── _browser.py (Playwright)      │                      │
-│  │  ├── _primitives.py (shell/http/   │                      │
+│  │  ├── browser.py (Playwright)       │                      │
+│  │  ├── primitives.py (shell/http/    │                      │
 │  │  │    note)                         │                      │
-│  │  ├── _dom.py (HTML → ≤2KB summary) │                      │
+│  │  ├── dom.py (HTML → ≤2KB summary)  │                      │
 │  │  ├── guardrails.py (sanitizer)     │                      │
-│  │  └── _rpc.py (UNIX socket server)  │                      │
+│  │  └── rpc.py (UNIX socket server)   │                      │
 │  └────────────────────────────────────┘                      │
 │           │                                                   │
 │           │ proxy: http://127.0.0.1:8081                     │
