+++ specs/FIELD-CRITIQUE-V2-ASSESSMENT.md
@@ -11,12 +11,12 @@
 | # | Critique | Verdict | Root Cause | Fix |
 |---|----------|---------|-----------|-----|
 | 1 | Nmap blocked (libcap Operation not permitted) | **True** — never tested | `no-new-privileges:true` in docker-compose.yml blocks raw sockets | Add `cap_add: [NET_RAW]` or default to `nmap -sT` (TCP connect) |
-| 10 | Browser can't detect XSS (no dialog capture) | **True** — MVP miss | `_browser.py` has no `page.on('dialog')` listener | ~20 lines: capture dialog message, auto-dismiss, include in turn response |
+| 10 | Browser can't detect XSS (no dialog capture) | **True** — MVP miss | `browser.py` has no `page.on('dialog')` listener | ~20 lines: capture dialog message, auto-dismiss, include in turn response |
 | 12 | Auth lost across turns (token expiry) | **Partially true** | Login flow works but no auto-refresh on 401. Tested initial login only, never tested expiry. | Detect 401 in http/browser response → re-run auth.yaml login → retry |
 | 2 | Pipelines sequential not parallel | **True** | `scanCmd()` runs one pipeline at a time, no goroutines | Allow comma-separated pipelines, run in parallel goroutines |
 | 16 | Note doesn't affect next-turn reasoning | **True** — oversight | Dispatcher returns turn metadata but doesn't include recent findings | Include last 5 findings in every turn response (~10 lines) |
 | 19 | WebSocket traffic not captured | **True** | `interceptor.py` only has `request()` and `response()` hooks | Add `websocket_message()` hook (~30 lines) |
-| 17 | DOM summary text-only (no JS errors) | **True** | `_browser.py` doesn't listen for console errors or network failures | Add `page.on('console')` and `page.on('pageerror')` listeners |
+| 17 | DOM summary text-only (no JS errors) | **True** | `browser.py` doesn't listen for console errors or network failures | Add `page.on('console')` and `page.on('pageerror')` listeners |
 | 5 | Fragile pipelines (silent failure) | **True** | Tools fail but pipeline continues with `|| true` — no explanation of WHY | Parse stderr for common error patterns, surface in structured output |
 
 **Bottom line: 3 small bugs (dialog capture, nmap cap, findings feedback) account for ~80% of the field test pain. All are <30 lines each.**
@@ -69,12 +69,12 @@ These are genuine misses that should have been caught before release:
 
 | What | How bad | Fix effort |
 |------|---------|-----------|
-| **No dialog capture** | Blocks all XSS verification — ~10 Juice Shop challenges need `alert()` | 20 lines in `_browser.py` |
+| **No dialog capture** | Blocks all XSS verification — ~10 Juice Shop challenges need `alert()` | 20 lines in `browser.py` |
 | **Nmap blocked** | Can't port scan from inside the container | 1 line in `docker-compose.yml` |
 | **Findings not in turn response** | Agent records findings but can't see them in next turn | 10 lines in Dispatcher |
 | **No WebSocket capture** | Invisible to interceptor on Socket.IO/WS apps | 30 lines in `interceptor.py` |
-| **No console error capture** | JS errors (from XSS payloads, broken imports) invisible to agent | 15 lines in `_browser.py` |
-| **Auth token expiry not handled** | Token expires mid-session, agent gets 401 without explanation | 30 lines in `_auth.py` + `_browser.py` |
+| **No console error capture** | JS errors (from XSS payloads, broken imports) invisible to agent | 15 lines in `browser.py` |
+| **Auth token expiry not handled** | Token expires mid-session, agent gets 401 without explanation | 30 lines in `auth.py` + `browser.py` |
 
 **Total: ~106 lines of code to fix all 6 issues.**
