+++ specs/FIELD-CRITIQUE-V2-FINAL.md
@@ -115,11 +115,11 @@ The 38-challenge gap (46% → 80%) is the black-box ceiling.
 
 ```
 Week 1 — P0 bug fixes (unblock real usage):
-  1. Dialog capture                       20 LoC    _browser.py
-  2. Playwright event hooks (all 5)       25 LoC    _browser.py
+  1. Dialog capture                       20 LoC    browser.py
+  2. Playwright event hooks (all 5)       25 LoC    browser.py
   3. Nmap cap fix                          5 LoC    docker-compose.yml
   4. Findings in turn response            10 LoC    loop.py Dispatcher
-  5. Auth auto-refresh on 401             30 LoC    _auth.py + _browser.py
+  5. Auth auto-refresh on 401             30 LoC    auth.py + browser.py
   6. Pipeline structured error feedback   30 LoC    main.go
 
 Week 2 — Pattern recipes (highest leverage docs):
