+++ specs/FIELD-CRITIQUE-V2.md
@@ -142,7 +142,7 @@ For challenges requiring external knowledge (security questions, geo-stalking, e
 ```
 Phase 1 (P0 — unblock real usage):
   - nmap cap fix (docker-compose.yml)
-  - dialog capture for XSS (_browser.py)
+  - dialog capture for XSS (browser.py)
   - auth token persistence across turns + pipelines
   - authenticated tool invocation (nuclei/dalfox/sqlmap with session)
