+++ tools/copilot/__init__.py
@@ -1,0 +1,1 @@
+# pentest-kit copilot — browser-driven, LLM-reasoned pentesting mode.
