+++ tools/copilot/auth.py
