+++ tools/copilot/browser.py
@@ -35,7 +35,7 @@ except ImportError:
     Page = Any  # type: ignore[assignment,misc]
     async_playwright = None  # type: ignore[assignment]
 
-from _dom import DomSummary, summarize
+from dom import DomSummary, summarize
 from guardrails import sanitize
 
 # Timeouts
@@ -396,7 +396,7 @@ class BrowserWrapper:
                 html = ""
 
             from guardrails import sanitize
-            from _dom import summarize
+            from dom import summarize
             sanitized = sanitize(html, logdir=self.copilot_dir, source_url=url)
             dom = summarize(sanitized.cleaned)
