+++ tools/copilot/loop.py
@@ -41,11 +41,11 @@ from typing import Any
 # Ensure the copilot tools directory is importable.
 sys.path.insert(0, str(Path(__file__).parent))
 
-from _auth import AuthManager, load_auth_config  # noqa: E402
-from _browser import BrowserWrapper  # noqa: E402
-from _primitives import PrimitiveError, run_http, run_note, run_shell  # noqa: E402
-from _rpc import make_server  # noqa: E402
-from _transcript import (  # noqa: E402
+from auth import AuthManager, load_auth_config  # noqa: E402
+from browser import BrowserWrapper  # noqa: E402
+from primitives import PrimitiveError, run_http, run_note, run_shell  # noqa: E402
+from rpc import make_server  # noqa: E402
+from transcript import (  # noqa: E402
     SummaryManager,
     StateManager,
     TranscriptWriter,
