+++ tools/copilot/pyproject.toml
@@ -1,0 +1,45 @@
+[tool.poetry]
+name = "pentest-kit-copilot"
+version = "0.12.1"
+description = "Browser-driven copilot mode for pentest-kit"
+authors = ["nunenuh"]
+license = "MIT"
+package-mode = false
+
+[tool.poetry.dependencies]
+python = "^3.10"
+pyyaml = ">=6.0"
+
+[tool.poetry.group.browser.dependencies]
+playwright = ">=1.40"
+
+[tool.poetry.group.proxy.dependencies]
+mitmproxy = ">=10.0"
+
+[tool.poetry.group.crypto.dependencies]
+pyzmq = ">=25.0"
+pycryptodomex = ">=3.19"
+Pillow = ">=10.0"
+pyotp = ">=2.9"
+
+[tool.poetry.group.dev.dependencies]
+pytest = ">=8.0"
+pytest-asyncio = ">=0.23"
+pytest-cov = ">=4.0"
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+asyncio_mode = "auto"
+pythonpath = ["."]
+
+[tool.coverage.run]
+source = ["."]
+omit = ["tests/*", "__init__.py"]
+
+[tool.coverage.report]
+show_missing = true
+fail_under = 60
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
