+++ tools/copilot/tests/conftest.py
@@ -1,0 +1,15 @@
+"""Shared fixtures for copilot tests."""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+import pytest
+
+
+@pytest.fixture
+def tmp_copilot(tmp_path: Path) -> Path:
+    """Create a temp copilot dir with standard subdirs."""
+    for subdir in ("screenshots", "interceptor", "sessions"):
+        (tmp_path / subdir).mkdir()
+    return tmp_path
