+++ tools/copilot/tests/test_auth.py
@@ -1,4 +1,4 @@
-"""Tests for _auth — config loading and credential management (no browser)."""
+"""Tests for auth — config loading and credential management (no browser)."""
 
 from __future__ import annotations
 
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from _auth import AuthConfig, AuthManager, Credential, load_auth_config
+from auth import AuthConfig, AuthManager, Credential, load_auth_config
 
 
 @pytest.fixture()
