+++ tools/copilot/tests/test_browser.py
@@ -1,4 +1,4 @@
-"""Tests for _browser.py — new actions: open_tab, http_parallel, set_extra_headers, dispatch."""
+"""Tests for browser.py — new actions: open_tab, http_parallel, set_extra_headers, dispatch."""
 
 from __future__ import annotations
 
@@ -24,7 +24,7 @@ def tmp_copilot(tmp_path: Path) -> Path:
 @pytest.fixture
 def wrapper(tmp_copilot: Path):
     """Create a BrowserWrapper with mocked internals."""
-    from _browser import BrowserWrapper
+    from browser import BrowserWrapper
 
     bw = BrowserWrapper(copilot_dir=tmp_copilot)
     bw._started = True
@@ -216,7 +216,7 @@ async def test_http_parallel_handles_errors(wrapper):
 
 def test_extra_headers_default(tmp_copilot: Path):
     """_extra_headers should default to empty dict."""
-    from _browser import BrowserWrapper
+    from browser import BrowserWrapper
 
     bw = BrowserWrapper(copilot_dir=tmp_copilot)
     assert bw._extra_headers == {}
