+++ tools/copilot/tests/test_dom.py
@@ -1,10 +1,10 @@
-"""Tests for _dom.summarize — no browser dependency."""
+"""Tests for dom.summarize — no browser dependency."""
 
 from __future__ import annotations
 
 import pytest
 
-from _dom import MAX_SUMMARY_BYTES, summarize
+from dom import MAX_SUMMARY_BYTES, summarize
 
 
 def test_empty_html():
