+++ tools/copilot/tests/test_primitives.py
@@ -1,4 +1,4 @@
-"""Tests for _primitives (shell, http, note) — no browser, no mitmproxy."""
+"""Tests for primitives (shell, http, note) — no browser, no mitmproxy."""
 
 from __future__ import annotations
 
@@ -9,7 +9,7 @@ from pathlib import Path
 
 import pytest
 
-from _primitives import (
+from primitives import (
     MAX_BODY_BYTES,
     PrimitiveError,
     run_http,
@@ -68,10 +68,10 @@ def test_shell_stdout_truncated_when_large():
 
 def test_shell_timeout(monkeypatch: pytest.MonkeyPatch):
     # Monkeypatch the timeout constant to 1s so the test runs fast.
-    import _primitives
-    monkeypatch.setattr(_primitives, "SHELL_TIMEOUT_SEC", 1)
+    import primitives
+    monkeypatch.setattr(primitives, "SHELL_TIMEOUT_SEC", 1)
     with pytest.raises(PrimitiveError) as exc:
-        _primitives.run_shell(["sleep", "5"])
+        primitives.run_shell(["sleep", "5"])
     assert exc.value.code == "timeout"
