+++ tools/copilot/tests/test_rpc.py
@@ -1,4 +1,4 @@
-"""Tests for _rpc — parse/dispatch layer, then end-to-end socket roundtrip."""
+"""Tests for rpc — parse/dispatch layer, then end-to-end socket roundtrip."""
 
 from __future__ import annotations
 
@@ -9,7 +9,7 @@ from typing import Any
 
 import pytest
 
-from _rpc import (
+from rpc import (
     RpcRequest,
     handle_line,
     make_server,
