+++ tools/copilot/tests/test_transcript.py
@@ -1,4 +1,4 @@
-"""Tests for _transcript — writer, summary, state, resume context."""
+"""Tests for transcript — writer, summary, state, resume context."""
 
 from __future__ import annotations
 
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-from _transcript import (
+from transcript import (
     RESUME_TAIL_TURNS,
     SUMMARY_INTERVAL,
     SummaryManager,
