+++ tools/copilot/transcript.py
